armadillo_get_number_of_omp_threads {RcppArmadillo}	R Documentation

Report (or Set) Maximum Number of OpenMP Threads

Description

Report (or Set) Maximum Number of OpenMP Threads

Usage

armadillo_get_number_of_omp_threads()

armadillo_set_number_of_omp_threads(n)

Arguments

n
Number of threads to be set

Value

For the getter, and on a system with OpenMP, the maximum number of threads that OpenMP may be using and on systems without it, one. The setter does not return a value.

[Package RcppArmadillo version 14.0.0-1 Index]