armadillo_set_seed {RcppArmadillo}	R Documentation

Set the Armadillo Random Number Generator to the given value

Description

Set the Armadillo Random Number Generator to the given value

Usage

armadillo_set_seed(val)

Arguments

val
The seed used to initialize Armadillo's random number generator.

Details

Depending on whether RcppArmadillo was compiled for the C++98 standard (currently the default) or for C++11 (optional), two different RNGs may be used. This function resets either. For C++98, the system library RNG is used. It may have poor performance, particularly on older versions of Windows. For C++11, the RNG included in its library is used.

Value

The function is invoked for its side effect and has no return value.

Note

This has been found to not work as espected in RStudio as its code also uses the system RNG library. You may have to either not run within RStudio or change your code to use a different RNG such as the one from R.

See Also

The R documentation on its RNGs all of which are accessible via Rcpp.

[Package RcppArmadillo version 0.4.650.1.1 Index]