armadillo_set_seed_random {RcppArmadillo}	R Documentation

Set the Armadillo Random Number Generator to a random value

Description

Set the Armadillo Random Number Generator to a random value

Usage


armadillo_set_seed_random()

Details

Depending on whether RcppArmadillo was compiled for the C++98 standard (currently the default) or for C++11 (optional), two different RNGs may be used. This function resets either. For C++98, the R programming language's RNG is used. For C++11, the RNG included in the <random> library is used only when #define ARMA_USE_CXX11_RNG is placed before #include <RcppArmadillo.h>. Otherwise, the R programming language's RNG will be used.

Value

The function is invoked for its side effect and has no return value.

Note

This has been found to not work as espected in RStudio as its code also uses the system RNG library. You may have to either not run within RStudio or change your code to use a different RNG such as the one from R.

See Also

The R documentation on its RNGs all of which are accessible via Rcpp.

[Package RcppArmadillo version 0.10.7.3.0 Index]