armadillo_throttle_cores {RcppArmadillo}	R Documentation

Throttle (or Reset) (Rcpp)Armadillo to Two Cores

Description

Helper functions to throttle use of cores by RcppArmadillo-internal code on systems with OpenMP. On package load, the initial value is saved and used to reset the value.

Usage

armadillo_throttle_cores(n = 2)

armadillo_reset_cores()

Arguments

n
Integer value of desired cores, default is two

[Package RcppArmadillo version 14.0.0-1 Index]