armadillo_version {RcppArmadillo}	R Documentation

Report the version of Armadillo

Description

Report the version of Armadillo

Usage


armadillo_version(single)

Arguments

single
A logical vector indicating whether a single return values is requested, or a named vector with three elements major, minor and patch.

Details

The version is defined by Armadillo in the header arma_version.hpp.

Value

Depending on the value of single, either a single number describing the Armadillo version or a named vector with three elements major, minor and patch.

See Also

Armadillo header file arma_version.hpp.

[Package RcppArmadillo version 0.10.7.3.0 Index]